pyproject.toml
[project]
name = "grun"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.6.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.7",
    "mypy>=1.16.1",
    "neo4j>=5.28.1",
    "pandas>=2.3.1",
    "pandas-stubs>=2.3.0.250703",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "ruff>=0.12.3",
    "streamlit>=1.46.1",
    "types-requests>=2.32.4.20250611",
]
